refactor(mutation-kill): extract round-runner + split loop module
run_for_file's 11-param signature and mutation_kill_loop.py's five bundled
responsibility clusters were both pre-existing debt flagged by
complexity-review/structure-review. Bundle the run-shaped inputs into a
RunContext dataclass and extract the per-round body into _score_round()
(scoring + stop-checks) and _run_round() (generate/insert/verify/commit),
cutting run_for_file to source_file/ctx/generate/max_rounds. Split
insertion mechanics into mutation_kill_insert.py and headless generation +
CLI into mutation_kill_headless.py, mirroring the existing wrapper/report
module split — mutation_kill_loop.py now stays scoped to config +
orchestration.

mutation_kill_loop_python.py's reused generic helpers (strip_code_fences,
resolve_model, claude_cli_available, CLAUDE_CLI) now import from
mutation_kill_headless, where they actually live post-split. The
__main__ guard aliases this module in sys.modules before importing
mutation_kill_headless back, so `python mutation_kill_loop.py --headless`
doesn't re-execute this file as a second, distinct module object.

Closes #1561
Closes #1562
Part of #1568
