+++ tests/scripts/test_mutation_kill_loop.py
@@ -7,11 +7,17 @@ dotnet / git / Stryker subprocess is mocked — no real .NET tooling runs — an
 every project name is generic. Fixtures write a minimal ``stryker-config.json``
 and a fixed ``mutation-report.json`` to a tmp_path repo so path derivation and
 survivor extraction are exercised without a live mutation run.
+
+Insertion mechanics moved to ``test_mutation_kill_insert.py`` and headless
+generation / CLI moved to ``test_mutation_kill_headless.py`` (#1562) — this
+file now covers only config, the scoped Stryker run, verify/commit
+subprocess wiring, and ``run_for_file`` orchestration.
 """
 
 from __future__ import annotations
 
 import json
+import subprocess
 import sys
 from pathlib import Path
 
@@ -28,6 +34,8 @@ SCRIPTS_DIR = (
 )
 sys.path.insert(0, str(SCRIPTS_DIR))
 
+import mutation_kill_headless  # module-split literal check below
+import mutation_kill_insert  # module-split literal check below
 import mutation_kill_loop as loop
 
 FORBIDDEN_LITERALS = ["Aci.Speedpay", "Controllers", "AwesomeAssertions", "Moq", "AutoFixture"]
@@ -189,330 +197,6 @@ def test_scoped_run_delegates_dotnet_root_and_sln_to_wrapper(
 
 
 # =============================================================================
-# Scenario: Bare-CLI default mode with no generator fails fast at startup
-# =============================================================================
-def test_bare_cli_no_generator_fails_fast_at_startup(
-    monkeypatch: pytest.MonkeyPatch, capsys
-):
-    # Guard: neither a scoped Stryker run nor any subprocess may fire.
-    def explode(*a, **k):
-        raise AssertionError("startup preflight must run before any subprocess")
-
-    monkeypatch.setattr(loop, "run_scoped_stryker", explode)
-    monkeypatch.setattr(loop.subprocess, "run", explode)
-
-    rc = loop.main(["--config", "stryker-config.json", "--file", "PaymentService.cs"])
-
-    assert rc != 0
-    err = capsys.readouterr().err
-    assert (
-        "no test generator available — invoke via the mutation-kill agent "
-        "or pass --headless"
-    ) in err
-
-
-def test_headless_flag_does_not_trip_the_no_generator_preflight(
-    monkeypatch: pytest.MonkeyPatch, capsys
-):
-    # --headless is accepted (Slice 3 wires it); it must NOT emit the
-    # no-generator message. With the CLI present, the next gate is the
-    # required-file-args check — assert that exact message is emitted, which
-    # positively proves the no-generator preflight was passed (not merely that
-    # the no-generator text is absent). Pin claude_cli_available so the emitted
-    # message is deterministic regardless of the host's PATH.
-    monkeypatch.setattr(loop, "claude_cli_available", lambda: True)
-
-    rc = loop.main(["--headless"])
-
-    err = capsys.readouterr().err
-    assert loop.NO_GENERATOR_MESSAGE not in err
-    assert "--headless requires --file, --test-file, and --source-path" in err
-    assert rc != 0
-
-
-# =============================================================================
-# Slice 3 — Optional --headless generation mode
-# =============================================================================
-# Scenario: Headless mode generates via the Claude CLI
-def test_headless_generator_invokes_claude_print_and_strips_fences(
-    monkeypatch: pytest.MonkeyPatch,
-):
-    captured: dict = {}
-
-    class _R:
-        returncode = 0
-        stdout = "```csharp\n[Test]\npublic void New_Case_KillsMutant() {}\n```"
-        stderr = ""
-
-    def fake_run(argv, **kwargs):
-        captured["argv"] = argv
-        return _R()
-
-    monkeypatch.setattr(loop.subprocess, "run", fake_run)
-
-    generate = loop.make_headless_generator("some-test-model")
-    survivors = [_mutant("Survived", "ArithmeticOperator", 10)]
-    out = generate(
-        "PaymentService.cs",
-        survivors,
-        "public class PaymentService {}",
-        "public class PaymentServiceTests { }",
-    )
-
-    argv = captured["argv"]
-    assert argv[0] == loop.CLAUDE_CLI
-    assert "--print" in argv
-    # --model carries the resolved model.
-    assert argv[argv.index("--model") + 1] == "some-test-model"
-    prompt = argv[-1]
-    # The existing test file is the pattern, and the survivor summary is present.
-    assert "PaymentServiceTests" in prompt
-    assert "ArithmeticOperator" in prompt
-    # No hardcoded library name leaks into the prompt.
-    assert not any(lit in prompt for lit in FORBIDDEN_LITERALS)
-    # Fences stripped from the returned methods.
-    assert "```" not in out
-    assert "New_Case_KillsMutant" in out
-
-
-# Scenario: --model resolves from the flag, then the env var, else None
-def test_model_resolves_from_env_when_set(monkeypatch: pytest.MonkeyPatch):
-    monkeypatch.setenv("DEV_TEAM_MUTATION_MODEL", "env-model")
-    assert loop.resolve_model() == "env-model"
-    # An explicit --model wins over the env var.
-    assert loop.resolve_model("flag-model") == "flag-model"
-
-
-def test_model_resolves_to_none_when_env_unset(
-    monkeypatch: pytest.MonkeyPatch,
-):
-    # No model snapshot id is pinned in source (cf. ADR 0008 / no-pinned-snapshots
-    # guard); unresolved means None so `claude --print` uses its own default.
-    monkeypatch.delenv("DEV_TEAM_MUTATION_MODEL", raising=False)
-    assert loop.resolve_model() is None
-
-
-def test_headless_omits_model_flag_when_unresolved(monkeypatch: pytest.MonkeyPatch):
-    captured = {}
-
-    class _R:
-        returncode = 0
-        stdout = "void New_Case() {}"
-        stderr = ""
-
-    def fake_run(argv, **k):
-        captured["argv"] = argv
-        return _R()
-
-    monkeypatch.setattr(loop.subprocess, "run", fake_run)
-    generate = loop.make_headless_generator(None)
-    generate("S.cs", [_mutant("Survived", "ArithmeticOperator", 10)], "class S {}", "class T {}")
-    # --model is absent entirely; claude --print falls back to its own default.
-    assert "--model" not in captured["argv"]
-    assert "--print" in captured["argv"]
-
-
-# Scenario: Default (non-headless) mode spawns no Claude subprocess
-def test_default_non_headless_spawns_no_claude_subprocess(
-    monkeypatch: pytest.MonkeyPatch, capsys
-):
-    def explode(*a, **k):
-        raise AssertionError("no subprocess may be spawned in the default mode")
-
-    monkeypatch.setattr(loop.subprocess, "run", explode)
-    monkeypatch.setattr(
-        loop, "claude_cli_available", lambda: pytest.fail("must not probe the CLI")
-    )
-
-    rc = loop.main(["--config", "stryker-config.json", "--file", "Foo.cs"])
-
-    assert rc != 0
-    assert loop.NO_GENERATOR_MESSAGE in capsys.readouterr().err
-
-
-# Scenario: Missing Claude CLI under headless fails cleanly and names the fix
-def test_missing_claude_cli_under_headless_names_remediation(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch, capsys
-):
-    monkeypatch.setattr(loop, "claude_cli_available", lambda: False)
-    # No file may be mutated: run_for_file must never be reached.
-    monkeypatch.setattr(
-        loop, "run_for_file", lambda *a, **k: pytest.fail("must not run — CLI missing")
-    )
-
-    rc = loop.main(
-        [
-            "--headless",
-            "--file", "Foo.cs",
-            "--test-file", str(tmp_path / "FooTests.cs"),
-            "--source-path", str(tmp_path / "Foo.cs"),
-        ]
-    )
-
-    assert rc != 0
-    err = capsys.readouterr().err
-    # Names how to install AND authenticate the CLI.
-    assert "install" in err.lower()
-    assert "claude" in err.lower()
-    assert "authenticate" in err.lower() or "ANTHROPIC_API_KEY" in err
-
-
-def test_claude_cli_available_reflects_subprocess_outcome(
-    monkeypatch: pytest.MonkeyPatch,
-):
-    class _OK:
-        returncode = 0
-
-    monkeypatch.setattr(loop.subprocess, "run", lambda *a, **k: _OK())
-    assert loop.claude_cli_available() is True
-
-    def _missing(*a, **k):
-        raise FileNotFoundError("claude")
-
-    monkeypatch.setattr(loop.subprocess, "run", _missing)
-    assert loop.claude_cli_available() is False
-
-
-# =============================================================================
-# Insert-mechanics fixtures
-# =============================================================================
-_BLOCK_NAMESPACE_CLASS = (
-    "namespace Widget.Tests\n"
-    "{\n"
-    "    public class PaymentServiceTests\n"
-    "    {\n"
-    "        [Test]\n"
-    "        public async Task Existing_Case_Works()\n"
-    "        {\n"
-    "        }\n"
-    "    }\n"
-    "}\n"
-)
-
-_FILE_SCOPED_CLASS = (
-    "namespace Widget.Tests;\n"
-    "\n"
-    "public class PaymentServiceTests\n"
-    "{\n"
-    "    [Test]\n"
-    "    public async Task Existing_Case_Works()\n"
-    "    {\n"
-    "    }\n"
-    "}\n"
-)
-
-_NEW_METHOD = (
-    "        [Test]\n"
-    "        public async Task New_Case_KillsMutant()\n"
-    "        {\n"
-    "        }\n"
-)
-
-
-# =============================================================================
-# Scenario: Duplicate method names abort insertion
-# =============================================================================
-def test_duplicate_method_names_abort_insertion(tmp_path: Path):
-    test_file = tmp_path / "PaymentServiceTests.cs"
-    test_file.write_text(_BLOCK_NAMESPACE_CLASS, encoding="utf-8")
-    before = test_file.read_text(encoding="utf-8")
-
-    dup_block = (
-        "        [Test]\n"
-        "        public async Task Existing_Case_Works()\n"
-        "        {\n"
-        "        }\n"
-    )
-
-    outcome = loop.apply_generated_methods(test_file, dup_block)
-
-    assert outcome.inserted is False
-    assert "duplicate" in outcome.reason.lower()
-    # File is unchanged.
-    assert test_file.read_text(encoding="utf-8") == before
-
-
-def test_detect_duplicate_methods_reports_only_collisions():
-    dupes = loop.detect_duplicate_methods(
-        _BLOCK_NAMESPACE_CLASS,
-        "public async Task Existing_Case_Works() {}\n"
-        "public async Task Brand_New_Case() {}\n",
-    )
-    assert dupes == ["Existing_Case_Works"]
-
-
-# =============================================================================
-# Scenario: Generated methods are inserted before the class-closing brace
-# =============================================================================
-def test_methods_inserted_before_class_closing_brace(tmp_path: Path):
-    test_file = tmp_path / "PaymentServiceTests.cs"
-    test_file.write_text(_BLOCK_NAMESPACE_CLASS, encoding="utf-8")
-
-    outcome = loop.apply_generated_methods(test_file, _NEW_METHOD)
-
-    assert outcome.inserted is True
-    lines = test_file.read_text(encoding="utf-8").splitlines()
-    new_idx = next(i for i, ln in enumerate(lines) if "New_Case_KillsMutant" in ln)
-    # The class-close ("    }") and namespace-close ("}") follow the new method.
-    class_close_idx = next(
-        i for i in range(new_idx + 1, len(lines)) if lines[i].rstrip("\r\n") == "    }"
-    )
-    ns_close_idx = next(
-        i for i in range(class_close_idx + 1, len(lines)) if lines[i].strip() == "}"
-    )
-    assert new_idx < class_close_idx < ns_close_idx
-
-
-# =============================================================================
-# Scenario: Insertion into a file-scoped-namespace class is handled or refused
-# =============================================================================
-def test_file_scoped_namespace_is_refused_not_corrupted(tmp_path: Path):
-    test_file = tmp_path / "PaymentServiceTests.cs"
-    test_file.write_text(_FILE_SCOPED_CLASS, encoding="utf-8")
-    before = test_file.read_text(encoding="utf-8")
-
-    with pytest.raises(loop.InsertionRefused) as exc:
-        loop.insert_before_class_close(test_file, _NEW_METHOD)
-
-    assert "file-scoped" in str(exc.value).lower()
-    # Never silently appended — file is byte-for-byte unchanged.
-    assert test_file.read_text(encoding="utf-8") == before
-
-
-def test_non_four_space_indent_is_refused(tmp_path: Path):
-    # Block namespace but the class body is tab-indented — the 4-space
-    # class-close heuristic must refuse rather than mis-insert.
-    tabbed = (
-        "namespace Widget.Tests\n"
-        "{\n"
-        "\tpublic class PaymentServiceTests\n"
-        "\t{\n"
-        "\t}\n"
-        "}\n"
-    )
-    test_file = tmp_path / "PaymentServiceTests.cs"
-    test_file.write_text(tabbed, encoding="utf-8")
-    before = test_file.read_text(encoding="utf-8")
-
-    with pytest.raises(loop.InsertionRefused):
-        loop.insert_before_class_close(test_file, _NEW_METHOD)
-
-    assert test_file.read_text(encoding="utf-8") == before
-
-
-def test_apply_wraps_refusal_as_outcome_without_writing(tmp_path: Path):
-    test_file = tmp_path / "PaymentServiceTests.cs"
-    test_file.write_text(_FILE_SCOPED_CLASS, encoding="utf-8")
-    before = test_file.read_text(encoding="utf-8")
-
-    outcome = loop.apply_generated_methods(test_file, _NEW_METHOD)
-
-    assert outcome.inserted is False
-    assert "file-scoped" in outcome.reason.lower()
-    assert test_file.read_text(encoding="utf-8") == before
-
-
-# =============================================================================
 # run_for_file harness — mocks every dotnet / git / Stryker touch.
 # =============================================================================
 def _loop_fixture(
@@ -521,18 +205,35 @@ def _loop_fixture(
     mutants=None,
     *,
     report_override: dict | None = None,
+    log=lambda m: None,
 ):
     """Wire a run_for_file invocation with a fixed initial report and a
-    generator that records its calls. Returns (kwargs, events).
+    generator that records its calls. Returns (source_file, ctx, kwargs, events).
 
     By default the initial report is the single-file shape ``_write_report``
     builds from ``mutants``. Pass ``report_override`` (a full report payload,
     e.g. a multi-file ``{"files": {...}}`` shape) instead when a caller needs
-    a custom report; ``mutants`` is ignored in that case.
+    a custom report; ``mutants`` is ignored in that case. Pass ``log`` to
+    capture round log lines (it lives on ``RunContext``, not the caller's
+    kwargs).
     """
     config = loop.load_loop_config(_write_config(tmp_path))
     test_file = tmp_path / "PaymentServiceTests.cs"
-    test_file.write_text(_BLOCK_NAMESPACE_CLASS, encoding="utf-8")
+    test_file.write_text(
+        (
+            "namespace Widget.Tests\n"
+            "{\n"
+            "    public class PaymentServiceTests\n"
+            "    {\n"
+            "        [Test]\n"
+            "        public async Task Existing_Case_Works()\n"
+            "        {\n"
+            "        }\n"
+            "    }\n"
+            "}\n"
+        ),
+        encoding="utf-8",
+    )
     source_path = tmp_path / "PaymentService.cs"
     source_path.write_text("public class PaymentService {}\n", encoding="utf-8")
     if report_override is not None:
@@ -544,27 +245,35 @@ def _loop_fixture(
 
     events: list = []
 
+    new_method = (
+        "        [Test]\n"
+        "        public async Task New_Case_KillsMutant()\n"
+        "        {\n"
+        "        }\n"
+    )
+
     def generator(src, survivors, src_text, test_text):
         events.append(("generate", len(survivors)))
-        return _NEW_METHOD
+        return new_method
 
     monkeypatch.setattr(loop, "git_revert", lambda tf, **k: events.append(("revert", str(tf))))
     monkeypatch.setattr(
         loop, "git_commit", lambda msg, tf, **k: events.append(("commit", msg)) or True
     )
 
+    ctx = loop.RunContext(
+        config=config,
+        test_file=test_file,
+        source_path=source_path,
+        output_dir=tmp_path / "out",
+        log=log,
+        initial_report_path=report,
+    )
     kwargs = {
-        "source_file": "PaymentService.cs",
-        "config": config,
-        "test_file": test_file,
-        "source_path": source_path,
-        "output_dir": tmp_path / "out",
         "generate": generator,
-        "initial_report_path": report,
         "max_rounds": 3,
-        "log": lambda m: None,
     }
-    return kwargs, events
+    return "PaymentService.cs", ctx, kwargs, events
 
 
 # =============================================================================
@@ -573,13 +282,13 @@ def _loop_fixture(
 def test_build_failure_after_insertion_is_reverted(
     tmp_path: Path, monkeypatch: pytest.MonkeyPatch
 ):
-    kwargs, events = _loop_fixture(tmp_path, monkeypatch, [_mutant("Survived")])
+    source_file, ctx, kwargs, events = _loop_fixture(tmp_path, monkeypatch, [_mutant("Survived")])
     monkeypatch.setattr(loop, "dotnet_build", lambda targets, **k: False)
     monkeypatch.setattr(
         loop, "dotnet_test", lambda *a, **k: pytest.fail("test must not run after build fail")
     )
 
-    loop.run_for_file(**kwargs)
+    loop.run_for_file(source_file, ctx, **kwargs)
 
     kinds = [e[0] for e in events]
     assert "revert" in kinds
@@ -592,11 +301,11 @@ def test_build_failure_after_insertion_is_reverted(
 def test_test_failure_after_insertion_is_reverted(
     tmp_path: Path, monkeypatch: pytest.MonkeyPatch
 ):
-    kwargs, events = _loop_fixture(tmp_path, monkeypatch, [_mutant("Survived")])
+    source_file, ctx, kwargs, events = _loop_fixture(tmp_path, monkeypatch, [_mutant("Survived")])
     monkeypatch.setattr(loop, "dotnet_build", lambda targets, **k: True)
     monkeypatch.setattr(loop, "dotnet_test", lambda targets, flt, **k: False)
 
-    loop.run_for_file(**kwargs)
+    loop.run_for_file(source_file, ctx, **kwargs)
 
     kinds = [e[0] for e in events]
     assert "revert" in kinds
@@ -607,7 +316,7 @@ def test_test_failure_after_insertion_is_reverted(
 # Green round commits (baseline for the revert scenarios above)
 # =============================================================================
 def test_green_round_commits(tmp_path: Path, monkeypatch: pytest.MonkeyPatch):
-    kwargs, events = _loop_fixture(tmp_path, monkeypatch, [_mutant("Survived")])
+    source_file, ctx, kwargs, events = _loop_fixture(tmp_path, monkeypatch, [_mutant("Survived")])
     monkeypatch.setattr(loop, "dotnet_build", lambda targets, **k: True)
     monkeypatch.setattr(loop, "dotnet_test", lambda targets, flt, **k: True)
     # Second round: scoped run returns a clean report so the loop stops "done".
@@ -615,7 +324,7 @@ def test_green_round_commits(tmp_path: Path, monkeypatch: pytest.MonkeyPatch):
     (tmp_path / "clean").mkdir(exist_ok=True)
     monkeypatch.setattr(loop, "run_scoped_stryker", lambda *a, **k: clean)
 
-    loop.run_for_file(**kwargs)
+    loop.run_for_file(source_file, ctx, **kwargs)
 
     kinds = [e[0] for e in events]
     assert "commit" in kinds
@@ -630,7 +339,7 @@ def test_non_improving_round_ends_the_file(
 ):
     # Round 1 (initial report) has 2 survivors; the scoped run for round 2
     # returns a report that still has 2 survivors — no improvement.
-    kwargs, events = _loop_fixture(
+    source_file, ctx, kwargs, events = _loop_fixture(
         tmp_path, monkeypatch, [_mutant("Survived", line=1), _mutant("Survived", line=2)]
     )
     monkeypatch.setattr(loop, "dotnet_build", lambda targets, **k: True)
@@ -643,7 +352,7 @@ def test_non_improving_round_ends_the_file(
     )
     monkeypatch.setattr(loop, "run_scoped_stryker", lambda *a, **k: stalled)
 
-    loop.run_for_file(**kwargs)
+    loop.run_for_file(source_file, ctx, **kwargs)
 
     # Exactly one commit (round 1); round 2 detected no improvement and stopped.
     commits = [e for e in events if e[0] == "commit"]
@@ -658,18 +367,20 @@ def test_round_log_line_uses_file_scoped_score_for_single_file_report(
 ):
     """A single-file scoped report — round log score must match that file's
     own counts (existing behavior, must stay equivalent)."""
-    kwargs, _events = _loop_fixture(
-        tmp_path, monkeypatch, [_mutant("Survived", line=1), _mutant("Killed", line=2)]
-    )
     logs: list[str] = []
-    kwargs["log"] = logs.append
+    source_file, ctx, kwargs, _events = _loop_fixture(
+        tmp_path,
+        monkeypatch,
+        [_mutant("Survived", line=1), _mutant("Killed", line=2)],
+        log=logs.append,
+    )
     monkeypatch.setattr(loop, "dotnet_build", lambda targets, **k: True)
     monkeypatch.setattr(loop, "dotnet_test", lambda targets, flt, **k: True)
     clean = _write_report(tmp_path / "clean", "src/Widget.WebApi/PaymentService.cs", [])
     (tmp_path / "clean").mkdir(exist_ok=True)
     monkeypatch.setattr(loop, "run_scoped_stryker", lambda *a, **k: clean)
 
-    loop.run_for_file(**kwargs)
+    loop.run_for_file(source_file, ctx, **kwargs)
 
     round_1_log = next(m for m in logs if m.startswith("  round 1:"))
     # 1 killed, 1 survived => honest = 1/2 * 100 = 50.0%
@@ -700,12 +411,11 @@ def test_round_log_line_scopes_to_target_file_in_multi_file_baseline_report(
             },
         }
     }
-    kwargs, _events = _loop_fixture(
-        tmp_path, monkeypatch, report_override=report_override
+    logs: list[str] = []
+    source_file, ctx, kwargs, _events = _loop_fixture(
+        tmp_path, monkeypatch, report_override=report_override, log=logs.append
     )
     kwargs["max_rounds"] = 2
-    logs: list[str] = []
-    kwargs["log"] = logs.append
     monkeypatch.setattr(loop, "dotnet_build", lambda targets, **k: True)
     monkeypatch.setattr(loop, "dotnet_test", lambda targets, flt, **k: True)
 
@@ -718,7 +428,7 @@ def test_round_log_line_scopes_to_target_file_in_multi_file_baseline_report(
     )
     monkeypatch.setattr(loop, "run_scoped_stryker", lambda *a, **k: round2_report)
 
-    loop.run_for_file(**kwargs)
+    loop.run_for_file(source_file, ctx, **kwargs)
 
     round_logs = [m for m in logs if m.startswith("  round")]
     assert len(round_logs) == 2, "expected both round 1 (baseline) and round 2 (scoped) to log"
@@ -745,9 +455,10 @@ def test_baseline_seeded_zero_survivors_never_calls_scoped_stryker(
             },
         }
     }
-    kwargs, events = _loop_fixture(tmp_path, monkeypatch, report_override=report_override)
     logs: list[str] = []
-    kwargs["log"] = logs.append
+    source_file, ctx, kwargs, events = _loop_fixture(
+        tmp_path, monkeypatch, report_override=report_override, log=logs.append
+    )
     monkeypatch.setattr(
         loop,
         "run_scoped_stryker",
@@ -757,7 +468,7 @@ def test_baseline_seeded_zero_survivors_never_calls_scoped_stryker(
         ),
     )
 
-    loop.run_for_file(**kwargs)
+    loop.run_for_file(source_file, ctx, **kwargs)
 
     assert any("no survivors" in msg for msg in logs)
     assert not any(e[0] in ("generate", "commit", "revert") for e in events)
@@ -811,10 +522,33 @@ def test_git_revert_checks_out_only_the_test_file(
 
 
 # =============================================================================
-# Scenario: The module carries no repo-specific literal
+# Scenario: `python mutation_kill_loop.py --headless ...` — the real
+# subprocess entry point stryker_shard_pipeline.py invokes by filename —
+# dispatches to mutation_kill_headless.main() without double-loading this
+# module (the sys.modules aliasing in the __main__ guard).
+# =============================================================================
+def test_script_invocation_dispatches_to_headless_main():
+    result = subprocess.run(
+        [sys.executable, str(SCRIPTS_DIR / "mutation_kill_loop.py"), "--file", "Foo.cs"],
+        capture_output=True,
+        text=True,
+        check=False,
+    )
+
+    assert result.returncode == 1
+    assert mutation_kill_headless.NO_GENERATOR_MESSAGE in result.stderr
+
+
+# =============================================================================
+# Scenario: No module in the mutation_kill_loop split carries a repo-specific
+# literal (#1562 split into three files — every one is checked).
 # =============================================================================
 def test_module_source_carries_no_repo_specific_literal():
-    source = (SCRIPTS_DIR / "mutation_kill_loop.py").read_text(encoding="utf-8")
-
-    present = [lit for lit in FORBIDDEN_LITERALS if lit in source]
-    assert present == [], f"repo-specific literals leaked into module: {present}"
+    for mod, filename in (
+        (loop, "mutation_kill_loop.py"),
+        (mutation_kill_insert, "mutation_kill_insert.py"),
+        (mutation_kill_headless, "mutation_kill_headless.py"),
+    ):
+        source = (SCRIPTS_DIR / filename).read_text(encoding="utf-8")
+        present = [lit for lit in FORBIDDEN_LITERALS if lit in source]
+        assert present == [], f"repo-specific literals leaked into {mod.__name__}: {present}"
